Add Inventory test class, factorize
